Split scanner runtime state helpers
